Remove TestPauseJobRecover
This test was mainly checking that tasks state is updated properly, especially
that tasks that were paused before killing scheduler are resumed on restart.

Since such a behaviour is already tested in SchedulerStateRecoverHelperTest, I
decided to remove the test.
